requirements/base.txt
Django
south
django-cms
django-countries
